Missing libwebp features

+++ src/aarch64/vp8pred_init.h
@@ -4,25 +4,25 @@
 #include "src/cpu.h"
 #include "src/vp8pred.h"
 
-void ff_pred16x16_vert_neon(uint8_t *, int);
-void ff_pred16x16_hor_neon(uint8_t *, int);
+void ff_pred16x16_vert_neon(uint8_t *, ptrdiff_t);
+void ff_pred16x16_hor_neon(uint8_t *, ptrdiff_t);
 
-void ff_pred4x4_tm_neon(uint8_t *, const uint8_t *, int);
-void ff_pred8x8_tm_neon(uint8_t *, int);
-void ff_pred16x16_tm_neon(uint8_t *, int);
-void ff_pred8x8_dc_neon(uint8_t *, int);
-void ff_pred16x16_dc_neon(uint8_t *, int);
+void ff_pred4x4_tm_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred8x8_tm_neon(uint8_t *, ptrdiff_t);
+void ff_pred16x16_tm_neon(uint8_t *, ptrdiff_t);
+void ff_pred8x8_dc_neon(uint8_t *, ptrdiff_t);
+void ff_pred16x16_dc_neon(uint8_t *, ptrdiff_t);
 
-void ff_pred4x4_dc_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_vert_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_hor_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_down_left_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_down_right_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_vert_left_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_vert_right_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_hor_up_neon(uint8_t *, const uint8_t *, int);
-void ff_pred4x4_hor_down_neon(uint8_t *, const uint8_t *, int);
-void ff_pred8x8_vert_neon(uint8_t *, int);
+void ff_pred4x4_dc_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_vert_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_hor_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_down_left_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_down_right_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_vert_left_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_vert_right_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_hor_up_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred4x4_hor_down_neon(uint8_t *, const uint8_t *, ptrdiff_t);
+void ff_pred8x8_vert_neon(uint8_t *, ptrdiff_t);
 
 static wpd_always_inline void ff_vp8_pred_init_aarch64(VP8PredContext *pred) {
     if (!(wpd_get_cpu_flags() & WPD_ARM_CPU_FLAG_NEON))
